test: merge DataCloudMetadataResultSetTest into MetadataResultSetsTest
DataCloudMetadataResultSet was deleted in this PR, but the test file
retained the old name and lived in the wrong package. Merge its empty-
result-set JDBC-shape smoke tests into the new MetadataResultSetsTest
under the .core.metadata package and delete the legacy file. No
behavior change.
